Tweet downloader implemented.
Got api keys for myself (anyone else will have to get their own, I'm not
checking my creds in!) Tested out the twitter4j library, which seems to
do the job, and do it well.

Implemented the target user timeline grab using a weird iterator thing
and then wrapped it in a stream; then set it up so it'll write out the
stuff for each user in the same format as the other data-grab.

Each user from the original set will be grabbed and their tweets will be
saved per-file.  Will do some counts and such in the morning once I have
the data.

Speaking of waiting, the iterator implements the wait-times so as to
avoid rate-limiting problems. Hopefully this works.

Also noticed that the java.util.Date class is perfect for dates. Good.
Gonna need that.