Fixed bugs involving loading and saving CSVs.
basically, had to stop using the scala package that I was using for CSV
and go to the library it was built on. meant building my own iterator
and stuff, but on the other hand it meant I could get quotes in tweets
escaped properly, allowing (freshly) downloaded tweets to be loaded
properly.

oh, also I had to re-run the downloader to get the new versions; in the
process I've changed the path stuff around a bunch. this also makes it
easier for the next stages, one of which is filtering/selecting a subset
of the downloaded tweets.