Implemented classifier, cleaned paths, added vec ops.
Classifier works, I think. It's a bit of a confusing piece of code, but
it does produce results, and it runs fast. The exact nature of those
results are stored in the resultnotes.md file, which is basically
preparation for my report.

Vector ops were added; they'll help with K-Means.

Also fixed everything up to start from a base path in the package
object. Much cleaner!